movie_review
The Movie Review project is a web application designed for users to explore and share their thoughts on various films. Built with Django and React, this platform allows users to submit their reviews, rate movies, and browse through a collection of reviews from other users. The project aims to create a community of movie enthusiasts who can engage in discussions and share their opinions on their favorite films.